Update YT to Audio.py
Added a few features:
The program will now check if the number of videos in the playlist is accurate to what YouTube reports (for example, privated videos still show up for playlist video count but aren't actually accessible). If a discrepancy exists, the program will notify the user and run as intended.

Discovered an issue where certain videos that remained on the list were not able to be downloaded. The program will now skip these & provide the urls for skipped videos after the program has finished running.